pdf-decryptor/requirements.txt
azure-functions==1.24.0
markupsafe==3.0.3
pip==24.0
pypdf==6.6.0
werkzeug==3.1.5
